Add cron-based UptimeService and connect to dev database script
This commit adds a UptimeService that conducts scheduled uptime checks using the robfig/cron library. Changes include initializing the cron scheduler and scheduling the periodic checks. A new method to stop the service is also added. In addition, the commit introduces a script to connect to the development database. Updates have been made to the project's dependencies and the Go files as well.